@NickLaMuro 2018-05-22
@miq-bot assign @gmcculloug 

@gmcculloug I realized you have been farming these out to some of the members of the automate team, but [this comment in particular](https://github.com/ManageIQ/manageiq/pull/17403/files#diff-1ce4751059050fe9933b971da4bb02ddR548) has me concerned, and I think you possibly had a hand in writing the original code.  Fine if you want to pass this off, but would like your input on if my thoughts on the `FIXME` were correct or misguided by something I am not accounting for.  Thanks.